third_party/blink/web_tests/external/wpt/css/cssom/HTMLLinkElement-disabled-001.tentative-expected.txt

This is a testharness.js-based test.
PASS <link disabled> prevents the stylesheet from being in document.styleSheets (from parser)
FAIL HTMLLinkElement.disabled reflects the <link disabled> attribute, and behaves consistently assert_equals: expected null but got Element node <link title="alt" rel="stylesheet" href="data:text/css,ht...
Harness: the test ran to completion.
